scan and fix code of middlewraes
