+++ src/api/middleWare/authMiddleWare/authInterfaces.ts
@@ -1,0 +1,10 @@
+import { Request } from "express";
+export interface tokenDataInterface{
+    email: string;
+    isAdmin: boolean;
+}
+
+export interface UserRequest extends Request {
+    user?: Partial<tokenDataInterface>;
+  }
+
