+++ src/api/middleWare/authMiddleWare/authService.ts
@@ -1,20 +1,14 @@
-import { tokenDataInterface, varifyToken } from "./Provider/jwt";
-import { handleError } from "../../utils/handleErrors";
-import { Request, Response, NextFunction } from "express";
+import { varifyToken } from "./jwt";
+import { handleError } from "../../../utils/handleErrors";
+import { Response, NextFunction } from "express";
+import { UserRequest, tokenDataInterface } from "./authInterfaces";
 
-export interface UserRequest extends Request {
-  user?: Partial<tokenDataInterface>;
-}
+
 const tokeGenerator = "jwt";
 const auth = (req: UserRequest, res: Response, next: NextFunction) => {
   if (tokeGenerator === "jwt") {
     try {
       const tokenFromClient = req.headers["access_token"];
-      // const cookies = req.cookies
-      // if (!cookies.access_token) res.status(401).json({message: "you are not provided cookie for token"})
-
-      // const tokenFromClient = cookies.access_token
-    //   console.log("tokenFromClient: ", tokenFromClient);
       if (typeof tokenFromClient === "string") {
         if (!tokenFromClient)
           throw new Error("Authentication error: please login");
